+++ tsconfig.json
@@ -4,8 +4,6 @@
     "declaration": true,
     "moduleResolution": "Bundler",
     "outDir": "./dist",
-    "rootDir": "./",
-    "allowImportingTsExtensions": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
